test: Migrate to jest
